refactor(safe-bash): extract regex allowlist module from permission-callback
Move SAFE_BASH_PATTERNS, isBashCommandSafe, and supporting regex constants
to a new server/safe-bash.ts sibling module, following the tool-tiers.ts /
tool-path-checker.ts / review-gate.ts extraction pattern. permission-callback.ts
re-exports the moved symbols so existing downstream consumers (cc-soleur-go
runner, batch-approval tests) continue working without import-path churn.

The near-miss telemetry WeakMap stays in permission-callback.ts because it is
keyed by CanUseToolContext (declared in that file); moving it would create a
cyclic import. permission-callback.ts drops from 840 to 677 lines (163-line
drop, AC2 target ≥80).

Ref #3040 (Finding 1 of 4).
